More build refactoring.
- test full 2.12, 2.13 and dotty matrix in CI
- update readme
- add help instructions on scalafmt --test failure
