+++ .scalafmt.conf
@@ -10,3 +10,9 @@ project.excludeFilters = [
   tests/macros/src/main/scala/scala/macros/tests/scaladays/main.scala,
   tests/macros/src/main/scala/scala/macros/tests/scaladays/Serialize.scala
 ]
+onTestFailure = """
+To fix this problem:
+1. run ./bin/scalafmt from the project root directory
+2. `git add . && git --amend` OR `git commit -am "Run scalafmt"`
+"""
+
