+++ bin/runci.sh
@@ -3,10 +3,8 @@ set -eux
 TEST=${1}
 
 case "$TEST" in
-  "cla" )
+  "check" )
     ./bin/checkCLA.sh
-    ;;
-  "scalafmt" )
     ./bin/scalafmt --test
     ;;
   * )
